// src/M4-SPRINT-6/S6-12-M4/src/schemas/address.schema.ts
import { z } from "zod";

export const addressSchema = z.object({
  id: z.number().positive(),
  street: z.string().max(45),
  zipCode: z.string().max(8),
  number: z.string().max(7).nullish(),
  city: z.string().max(20),
  state: z.string().max(2),
});

export const addressCreateSchema = addressSchema.omit({ id: true });